Virtual_TRy-On-master/requirements.txt
absl-py==2.1.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cvzone==1.6.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
gunicorn==23.0.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
onnxruntime==1.20.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.5
pycparser==2.22
PyMatting==1.1.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
referencing==0.35.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.0
sentencepiece==0.2.0
setuptools==75.1.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.3
tifffile==2024.12.12
tqdm==4.67.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
